Various Customer states + minor cleanup
Did a few of the customer states, I have a bit more in the works but is
full of errors so I'll add it when it's finished. Tried to stay
consistent with the formatting, if it's wrong I apologize and you can
change whatever I wrote. More coming in the morning.